test: split test_event_comments.py to fix 500-line CI limit
Move TestReactionToggle out to test_event_comment_reactions.py.
test_event_comments.py was 562 lines, over the agent-complexity
hard limit.
